+++ InvenTree/web/src/src/App.tsx
@@ -1,17 +1,9 @@
-import { t } from '@lingui/macro';
-import {
-  ColorScheme, ColorSchemeProvider, MantineProvider, MantineThemeOverride
-} from '@mantine/core';
-import { useColorScheme, useLocalStorage } from '@mantine/hooks';
-import { ModalsProvider } from '@mantine/modals';
-import { NotificationsProvider } from '@mantine/notifications';
 import * as Sentry from '@sentry/react';
 import { BrowserTracing } from '@sentry/tracing';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import axios from 'axios';
 import { useState } from 'react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { QrCodeModal } from './components/modals/QrCodeModal';
 import { useApiState } from './context/ApiState';
 import { AuthProvider } from './context/AuthContext';
 import { useLocalState } from './context/LocalState';
@@ -26,6 +18,7 @@ import { Profile } from './pages/Index/Profile/Profile';
 import Layout from './pages/layout';
 import { Login } from './pages/Login';
 import { Logout } from './pages/Logout';
+import { ThemeContext } from './ThemeContext';
 import { LanguageContext } from './translation';
 
 // Error tracking
@@ -96,50 +89,6 @@ const router = createBrowserRouter([
     errorElement: <ErrorPage />
   }
 ]);
-
-function ThemeContext({ children }: { children: any }) {
-  const [primaryColor, whiteColor, blackColor, radius, loader] = useLocalState((state) => [state.primaryColor, state.whiteColor, state.blackColor, state.radius, state.loader]);
-
-  // Color Scheme
-  const preferredColorScheme = useColorScheme();
-  const [colorScheme, setColorScheme] = useLocalStorage<ColorScheme>({
-    key: 'scheme',
-    defaultValue: preferredColorScheme
-  });
-  const toggleColorScheme = (value?: ColorScheme) => {
-    setColorScheme(value || (colorScheme === 'dark' ? 'light' : 'dark'));
-    myTheme.colorScheme = colorScheme;
-  };
-  const myTheme: MantineThemeOverride = {
-    colorScheme: colorScheme,
-    primaryColor: primaryColor,
-    white: whiteColor,
-    black: blackColor,
-    loader: loader,
-    defaultRadius: radius,
-  };
-
-  return (
-    <ColorSchemeProvider
-      colorScheme={colorScheme}
-      toggleColorScheme={toggleColorScheme}
-    >
-      <MantineProvider
-        theme={myTheme}
-        withGlobalStyles
-        withNormalizeCSS
-      >
-        <NotificationsProvider>
-          <ModalsProvider labels={{ confirm: t`Submit`, cancel: t`Cancel` }} modals={{ qr: QrCodeModal }}>
-
-            {children}
-
-          </ModalsProvider>
-        </NotificationsProvider>
-      </MantineProvider>
-    </ColorSchemeProvider>
-  )
-}
 
 // Main App
 export default function App() {
